// sdks/typescript/src/generated/Client.ts
export declare namespace StarlingClient {
    export type Options = BaseClientOptions;

    export interface RequestOptions extends BaseRequestOptions {}
}

export class StarlingClient {
    protected readonly _options: NormalizedClientOptionsWithAuth<StarlingClient.Options>;
    protected _accounts: AccountsClient | undefined;
    protected _business: BusinessClient | undefined;
    protected _individual: IndividualClient | undefined;
    protected _jointAccount: JointAccountClient | undefined;
    protected _soleTrader: SoleTraderClient | undefined;
    protected _accountHolder: AccountHolderClient | undefined;
    protected _profileImage: ProfileImageClient | undefined;
    protected _addresses: AddressesClient | undefined;
    protected _cards: CardsClient | undefined;
    protected _directDebits: DirectDebitsClient | undefined;
    protected _feed: FeedClient | undefined;
    protected _roundUp: RoundUpClient | undefined;
    protected _identity: IdentityClient | undefined;
    protected _payments: PaymentsClient | undefined;
    protected _payees: PayeesClient | undefined;
    protected _savingsGoals: SavingsGoalsClient | undefined;
    protected _settleUp: SettleUpClient | undefined;
    protected _spaces: SpacesClient | undefined;

    constructor(options: StarlingClient.Options = {}) {
        this._options = normalizeClientOptionsWithAuth(options);
    }

    public get accounts(): AccountsClient {
        return (this._accounts ??= new AccountsClient(this._options));
    }

    public get business(): BusinessClient {
        return (this._business ??= new BusinessClient(this._options));
    }

    public get individual(): IndividualClient {
        return (this._individual ??= new IndividualClient(this._options));
    }

    public get jointAccount(): JointAccountClient {
        return (this._jointAccount ??= new JointAccountClient(this._options));
    }

    public get soleTrader(): SoleTraderClient {
        return (this._soleTrader ??= new SoleTraderClient(this._options));
    }

    public get accountHolder(): AccountHolderClient {
        return (this._accountHolder ??= new AccountHolderClient(this._options));
    }

    public get profileImage(): ProfileImageClient {
        return (this._profileImage ??= new ProfileImageClient(this._options));
    }

    public get addresses(): AddressesClient {
        return (this._addresses ??= new AddressesClient(this._options));
    }

    public get cards(): CardsClient {
        return (this._cards ??= new CardsClient(this._options));
    }

    public get directDebits(): DirectDebitsClient {
        return (this._directDebits ??= new DirectDebitsClient(this._options));
    }

    public get feed(): FeedClient {
        return (this._feed ??= new FeedClient(this._options));
    }

    public get roundUp(): RoundUpClient {
        return (this._roundUp ??= new RoundUpClient(this._options));
    }

    public get identity(): IdentityClient {
        return (this._identity ??= new IdentityClient(this._options));
    }

    public get payments(): PaymentsClient {
        return (this._payments ??= new PaymentsClient(this._options));
    }

    public get payees(): PayeesClient {
        return (this._payees ??= new PayeesClient(this._options));
    }

    public get savingsGoals(): SavingsGoalsClient {
        return (this._savingsGoals ??= new SavingsGoalsClient(this._options));
    }

    public get settleUp(): SettleUpClient {
        return (this._settleUp ??= new SettleUpClient(this._options));
    }

    public get spaces(): SpacesClient {
        return (this._spaces ??= new SpacesClient(this._options));
    }

    /**
     * Make a passthrough request using the SDK's configured auth, retry, logging, etc.
     * This is useful for making requests to endpoints not yet supported in the SDK.
     * The input can be a URL string, URL object, or Request object. Relative paths are resolved against the configured base URL.
     *
     * @param {Request | string | URL} input - The URL, path, or Request object.
     * @param {RequestInit} init - Standard fetch RequestInit options.
     * @param {core.PassthroughRequest.RequestOptions} requestOptions - Per-request overrides (timeout, retries, headers, abort signal).
     * @returns {Promise<Response>} A standard Response object.
     */
    public async fetch(
        input: Request | string | URL,
        init?: RequestInit,
        requestOptions?: core.PassthroughRequest.RequestOptions,
    ): Promise<Response> {
        return core.makePassthroughRequest(
            input,
            init,
            {
                baseUrl: this._options.baseUrl ?? this._options.environment,
                headers: this._options.headers,
                timeoutInSeconds: this._options.timeoutInSeconds,
                maxRetries: this._options.maxRetries,
                fetch: this._options.fetch,
                logging: this._options.logging,
                getAuthHeaders: async () => (await this._options.authProvider.getAuthRequest()).headers,
            },
            requestOptions,
        );
    }
}