// sdks/typescript/src/generated/environments.ts
// This file was auto-generated by Fern from our API Definition.

export const StarlingEnvironment = {
    Sandbox: "https://api-sandbox.starlingbank.com",
    Production: "https://api.starlingbank.com",
} as const;

export type StarlingEnvironment = typeof StarlingEnvironment.Sandbox | typeof StarlingEnvironment.Production;